refactor: consolidate game websocket api structure
